Cambios, solicitud tipo carrito

+++ src/Inventory.Application/Dtos/ConteoDto.cs
@@ -3,12 +3,12 @@ namespace Inventory.Application.Dtos;
 public record ConteoDto(
     int Id, string SesionConteo, int ProductoId, string ProductoNombre,
     int UbicacionId, string UbicacionCodigo, int NumeroConteo,
-    decimal CantidadContada, int ContadoPorId, string ContadoPor, DateTime FechaConteo,
-    decimal ExistenciaSistema, decimal Diferencia
+    int CantidadContada, int ContadoPorId, string ContadoPor, DateTime FechaConteo,
+    int ExistenciaSistema, int Diferencia
 );
 
 public record RegistrarConteoDto(
-    string SesionConteo, int ProductoId, int UbicacionId, int NumeroConteo, decimal CantidadContada
+    string SesionConteo, int ProductoId, int UbicacionId, int NumeroConteo, int CantidadContada
 );
 
 // Sin UbicacionId a propósito (2026-09-15, pedido del operario): el operario elige
